Switched to private implementation using magic_enum
