+++ GridKit/Model/Evaluator.hpp
@@ -5,7 +5,7 @@
 #include <GridKit/CommonMath.hpp>
 #include <GridKit/Constants.hpp>
 #include <GridKit/LinearAlgebra/SparseMatrix/COO_Matrix.hpp>
-#include <GridKit/Model/VariableMonitorBase.hpp>
+#include <GridKit/Model/VariableMonitor.hpp>
 #include <GridKit/ScalarTraits.hpp>
 
 namespace GridKit
