+++ GridKit/Model/PhasorDynamics/Branch/BranchImpl.hpp
@@ -13,6 +13,7 @@
 #include <GridKit/Model/PhasorDynamics/Branch/Branch.hpp>
 #include <GridKit/Model/PhasorDynamics/Branch/BranchData.hpp>
 #include <GridKit/Model/PhasorDynamics/Bus/Bus.hpp>
+#include <GridKit/Model/VariableMonitorImpl.hpp>
 
 namespace GridKit
 {
@@ -75,7 +76,7 @@ namespace GridKit
     Branch<ScalarT, IdxT>::Branch(bus_type* bus1, bus_type* bus2, const model_data_type& data)
       : bus1_(bus1),
         bus2_(bus2),
-        monitor_(data)
+        monitor_(std::make_unique<MonitorT>(data))
     {
       initializeParameters(data);
       initializeMonitor();
@@ -286,23 +287,29 @@ namespace GridKit
     }
 
     template <class ScalarT, typename IdxT>
+    const Model::VariableMonitorBase* Branch<ScalarT, IdxT>::getMonitor() const
+    {
+      return monitor_.get();
+    }
+
+    template <class ScalarT, typename IdxT>
     void Branch<ScalarT, IdxT>::initializeMonitor()
     {
       using Variable = typename model_data_type::MonitorableVariables;
-      monitor_.set(Variable::ir1, [this]
-                   { return Ir1(); });
-      monitor_.set(Variable::ii1, [this]
-                   { return Ii1(); });
-      // monitor_.set(Variable::im1, [this] { return ?(); });
-      // monitor_.set(Variable::p1, [this] { return ?(); });
-      // monitor_.set(Variable::q1, [this] { return ?(); });
-      monitor_.set(Variable::ir2, [this]
-                   { return Ir2(); });
-      monitor_.set(Variable::ii2, [this]
-                   { return Ii2(); });
-      // monitor_.set(Variable::im2, [this] { return ?(); });
-      // monitor_.set(Variable::p2, [this] { return ?(); });
-      // monitor_.set(Variable::q2, [this] { return ?(); });
+      monitor_->set(Variable::ir1, [this]
+                    { return Ir1(); });
+      monitor_->set(Variable::ii1, [this]
+                    { return Ii1(); });
+      // monitor_->set(Variable::im1, [this] { return ?(); });
+      // monitor_->set(Variable::p1, [this] { return ?(); });
+      // monitor_->set(Variable::q1, [this] { return ?(); });
+      monitor_->set(Variable::ir2, [this]
+                    { return Ir2(); });
+      monitor_->set(Variable::ii2, [this]
+                    { return Ii2(); });
+      // monitor_->set(Variable::im2, [this] { return ?(); });
+      // monitor_->set(Variable::p2, [this] { return ?(); });
+      // monitor_->set(Variable::q2, [this] { return ?(); });
     }
 
     /**
