+++ GridKit/Model/PhasorDynamics/Branch/CMakeLists.txt
@@ -15,6 +15,8 @@ if(GRIDKIT_ENABLE_ENZYME)
       BranchEnzyme.cpp
     HEADERS
       ${_install_headers}
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include
     LINK_LIBRARIES
       PRIVATE ClangEnzymeFlags
     COMPILE_OPTIONS
@@ -24,11 +26,14 @@ else()
     SOURCES
       Branch.cpp
     HEADERS
-      ${_install_headers})
+      ${_install_headers}
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include)
 endif()
 
 gridkit_add_library(phasor_dynamics_branch_dependency_tracking
-  SOURCES BranchDependencyTracking.cpp)
+  SOURCES BranchDependencyTracking.cpp
+  INCLUDE_DIRECTORIES PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include)
 
 # Link to interface target for all components
 target_link_libraries(phasor_dynamics_components
