+++ GridKit/Model/PhasorDynamics/Bus/CMakeLists.txt
@@ -17,6 +17,8 @@ if(GRIDKIT_ENABLE_ENZYME)
       BusInfinite.cpp
     HEADERS
       ${_install_headers}
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include
     LINK_LIBRARIES
       PUBLIC GridKit::phasor_dynamics_core)
 else()
@@ -26,6 +28,8 @@ else()
       BusInfinite.cpp
     HEADERS
       ${_install_headers}
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include
     LINK_LIBRARIES
       PUBLIC GridKit::phasor_dynamics_core)
 endif()
@@ -34,6 +38,8 @@ gridkit_add_library(phasor_dynamics_bus_dependency_tracking
   SOURCES
     BusDependencyTracking.cpp
     BusInfiniteDependencyTracking.cpp
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include
   LINK_LIBRARIES
     PUBLIC GridKit::phasor_dynamics_core)
