+++ GridKit/Model/PhasorDynamics/BusBaseImpl.hpp
@@ -1,0 +1,33 @@
+#pragma once
+
+#include <GridKit/Model/PhasorDynamics/BusBase.hpp>
+#include <GridKit/Model/VariableMonitorImpl.hpp>
+
+namespace GridKit
+{
+  namespace PhasorDynamics
+  {
+    template <typename ScalarT, typename IdxT>
+    BusBase<ScalarT, IdxT>::BusBase(const BusData<RealT, IdxT>& data)
+      : bus_id_(data.bus_id),
+        monitor_(std::make_unique<MonitorT>("Bus_" + data.name, data.monitored_variables))
+    {
+      using Variable = typename BusData<RealT, IdxT>::MonitorableVariables;
+      monitor_->set(Variable::Vr, [this]
+                    { return Vr(); });
+      monitor_->set(Variable::Vi, [this]
+                    { return Vi(); });
+    }
+
+    template <typename ScalarT, typename IdxT>
+    BusBase<ScalarT, IdxT>::~BusBase()
+    {
+    }
+
+    template <typename ScalarT, typename IdxT>
+    inline const Model::VariableMonitorBase* BusBase<ScalarT, IdxT>::getMonitor() const
+    {
+      return monitor_.get();
+    }
+  } // namespace PhasorDynamics
+} // namespace GridKit
