+++ GridKit/Model/PhasorDynamics/BusFault/CMakeLists.txt
@@ -9,6 +9,8 @@ if(GRIDKIT_ENABLE_ENZYME)
       BusFaultEnzyme.cpp
     HEADERS
       ${_install_headers}
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include
     LINK_LIBRARIES
       PRIVATE ClangEnzymeFlags
     COMPILE_OPTIONS
@@ -18,11 +20,14 @@ else()
     SOURCES
       BusFault.cpp
     HEADERS
-      ${_install_headers})
+      ${_install_headers}
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include)
 endif()
 
 gridkit_add_library(phasor_dynamics_bus_fault_dependency_tracking
-  SOURCES BusFaultDependencyTracking.cpp)
+  SOURCES BusFaultDependencyTracking.cpp
+  INCLUDE_DIRECTORIES PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include)
 
 # Link to interface target for all components
 target_link_libraries(phasor_dynamics_components
