+++ GridKit/Model/PhasorDynamics/Exciter/IEEET1/CMakeLists.txt
@@ -13,6 +13,8 @@ gridkit_add_library(phasor_dynamics_exciter_ieeet1
     Ieeet1.cpp
   HEADERS
     ${_install_headers}
+  INCLUDE_DIRECTORIES
+    PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include
   LINK_LIBRARIES
     GridKit::phasor_dynamics_core
     GridKit::phasor_dynamics_signal)
@@ -20,6 +22,8 @@ gridkit_add_library(phasor_dynamics_exciter_ieeet1
 gridkit_add_library(phasor_dynamics_exciter_ieeet1_dependency_tracking
   SOURCES
     Ieeet1DependencyTracking.cpp
+  INCLUDE_DIRECTORIES
+    PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include
   LINK_LIBRARIES
     GridKit::phasor_dynamics_core
     GridKit::phasor_dynamics_signal_dependency_tracking)
