+++ GridKit/Model/PhasorDynamics/Exciter/IEEET1/Ieeet1Data.hpp
@@ -8,7 +8,6 @@
 #pragma once
 
 #include <GridKit/Model/PhasorDynamics/ComponentData.hpp>
-#include <GridKit/Utilities/Enum.hpp>
 
 namespace GridKit
 {
@@ -47,18 +46,8 @@ namespace GridKit
       enum class Ieeet1MonitorableVariables
       {
         efd,
-        ksat,
-        SIZE
+        ksat
       };
-
-      /**
-       * @brief Convert enum value to string label
-       */
-      inline const std::string& enumLabel(Ieeet1MonitorableVariables v)
-      {
-        static std::string labels[] = {"efd", "ksat"};
-        return labels[Utilities::enumId(v)];
-      }
 
       /**
        * @brief Contains modeling data for a IEEET1 Exciter model.
