+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENROUwS/GenrouData.hpp
@@ -7,7 +7,6 @@
 #pragma once
 
 #include <GridKit/Model/PhasorDynamics/ComponentData.hpp>
-#include <GridKit/Utilities/Enum.hpp>
 
 namespace GridKit
 {
@@ -54,24 +53,8 @@ namespace GridKit
       p,
       q,
       delta,
-      omega,
-      SIZE
+      omega
     };
-
-    /**
-     * @brief Convert enum value to string label
-     */
-    inline const std::string& enumLabel(GenrouMonitorableVariables v)
-    {
-      static const std::string labels[] = {
-          "ir",
-          "ii",
-          "p",
-          "q",
-          "delta",
-          "omega"};
-      return labels[Utilities::enumId(v)];
-    }
 
     /**
      * @brief Contains modeling data for a Genrou generator model.
