+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GenClassical/CMakeLists.txt
@@ -14,6 +14,8 @@ if(GRIDKIT_ENABLE_ENZYME)
       GenClassicalEnzyme.cpp
     HEADERS
       ${_install_headers}
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include
     LINK_LIBRARIES
       PRIVATE ClangEnzymeFlags
     COMPILE_OPTIONS
@@ -23,11 +25,14 @@ else()
     SOURCES
       GenClassical.cpp
     HEADERS
-      ${_install_headers})
+      ${_install_headers}
+    INCLUDE_DIRECTORIES
+      PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include)
 endif()
 
-  gridkit_add_library(phasor_dynamics_gen_classical_dependency_tracking
-    SOURCES GenClassicalDependencyTracking.cpp)
+gridkit_add_library(phasor_dynamics_gen_classical_dependency_tracking
+  SOURCES GenClassicalDependencyTracking.cpp
+  INCLUDE_DIRECTORIES PRIVATE ${GRIDKIT_THIRD_PARTY_DIR}/magic-enum/include)
 
 # Link to interface target for all components
 target_link_libraries(phasor_dynamics_components
