+++ GridKit/Utilities/CMakeLists.txt
@@ -11,7 +11,6 @@ add_subdirectory(CliOptions)
 install(
   FILES
     Colors.hpp
-    Enum.hpp
     FileIO.hpp
     MapFromCOO.hpp
   DESTINATION
